consolidated cruft test files into single files; lost some test as coverage is at 87% for now
