+++ .gitignore
@@ -16,9 +16,26 @@ parts/
 sdist/
 var/
 wheels/
+*.egg-info
 *.egg-info/
 .installed.cfg
 *.egg
+
+# Testing
+.coverage
+.coverage.*
+htmlcov/
+.pytest_cache/
+.tox/
+.nox/
+.hypothesis/
+coverage.json
+
+# Environment files
+shell.env
+devcontainer-environment-variables.json
+aws-profile-map.json
+.devcontainer/aws-profile-map.json
 
 # Virtual environments
 venv/
@@ -31,12 +48,11 @@ env/
 .vscode/
 *.swp
 *.swo
+*~
 
-# Devcontainer files
-shell.env
-devcontainer-environment-variables.json
-.devcontainer/aws-profile-map.json
+# OS files
+.DS_Store
+Thumbs.db
 
 # Test files
 output_file.sh
-coverage.json
