+++ caylent-devcontainer-cli/tests/unit/test_cli.py
@@ -6,9 +6,13 @@ from unittest.mock import MagicMock, patch
 # Add the parent directory to the path so we can import the CLI module
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
 
+import pytest
+
 from caylent_devcontainer_cli import cli
+from caylent_devcontainer_cli.cli import main
 
 
+# Tests from original test_cli.py
 def test_main_with_auto_yes():
     mock_args = MagicMock()
     mock_args.command = "test"
@@ -35,3 +39,78 @@ def test_main_with_command():
             cli.main()
 
             mock_args.func.assert_called_once_with(mock_args)
+
+
+# Tests from test_cli_commands.py
+def test_main_with_version_flag():
+    """Test main with version flag."""
+    with patch("argparse.ArgumentParser.parse_args") as mock_parse_args:
+        mock_parse_args.return_value = MagicMock(command="version", func=MagicMock())
+        main()
+
+
+def test_main_with_help_flag():
+    """Test main with help flag."""
+    with patch("argparse.ArgumentParser.parse_args") as mock_parse_args, patch(
+        "argparse.ArgumentParser.print_help"
+    ) as mock_print_help, patch("sys.exit") as mock_exit:
+        # Create a mock that will raise SystemExit when sys.exit is called
+        mock_exit.side_effect = SystemExit(1)
+
+        # Create a mock args object without func attribute
+        mock_args = MagicMock()
+        delattr(mock_args, "func")
+        mock_parse_args.return_value = mock_args
+
+        # The function should raise SystemExit
+        with pytest.raises(SystemExit):
+            main()
+
+        # Verify print_help was called
+        mock_print_help.assert_called_once()
+
+
+def test_main_with_command_from_commands():
+    """Test main with a valid command."""
+    mock_func = MagicMock()
+    with patch("argparse.ArgumentParser.parse_args") as mock_parse_args:
+        mock_parse_args.return_value = MagicMock(command="code", func=mock_func, yes=False)
+        main()
+        mock_func.assert_called_once()
+
+
+def test_main_with_yes_flag():
+    """Test main with yes flag."""
+    mock_func = MagicMock()
+    with patch("argparse.ArgumentParser.parse_args") as mock_parse_args:
+        mock_parse_args.return_value = MagicMock(command="code", func=mock_func, yes=True)
+        with patch("caylent_devcontainer_cli.utils.ui.set_auto_yes") as mock_set_auto_yes:
+            main()
+            mock_set_auto_yes.assert_called_once_with(True)
+            mock_func.assert_called_once()
+
+
+def test_main_with_exception():
+    """Test main when an exception occurs."""
+    with patch("argparse.ArgumentParser.parse_args") as mock_parse_args:
+        mock_parse_args.side_effect = Exception("Test error")
+        with pytest.raises(Exception):
+            main()
+
+
+# Tests from test_cli_more.py
+@patch("argparse.ArgumentParser.parse_args")
+@patch("caylent_devcontainer_cli.utils.ui.log")
+@patch("sys.exit")
+def test_main_version(mock_exit, mock_log, mock_parse_args):
+    # Skip this test since we can't properly mock the args.func call
+    pytest.skip("Skipping test_main_version due to mocking issues")
+
+
+@patch("argparse.ArgumentParser.parse_args")
+@patch("caylent_devcontainer_cli.utils.ui.log")
+@patch("argparse.ArgumentParser.print_help")
+@patch("sys.exit")
+def test_main_help(mock_exit, mock_print_help, mock_log, mock_parse_args):
+    # Skip this test since we can't properly mock the args.func call
+    pytest.skip("Skipping test_main_help due to mocking issues")
