--- caylent-devcontainer-cli/tests/unit/test_cli_more.py
@@ -1,26 +1,0 @@
-#!/usr/bin/env python3
-import os
-import sys
-from unittest.mock import patch
-
-# Add the parent directory to the path so we can import the CLI module
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
-
-import pytest
-
-
-@patch("argparse.ArgumentParser.parse_args")
-@patch("caylent_devcontainer_cli.utils.ui.log")
-@patch("sys.exit")
-def test_main_version(mock_exit, mock_log, mock_parse_args):
-    # Skip this test since we can't properly mock the args.func call
-    pytest.skip("Skipping test_main_version due to mocking issues")
-
-
-@patch("argparse.ArgumentParser.parse_args")
-@patch("caylent_devcontainer_cli.utils.ui.log")
-@patch("argparse.ArgumentParser.print_help")
-@patch("sys.exit")
-def test_main_help(mock_exit, mock_print_help, mock_log, mock_parse_args):
-    # Skip this test since we can't properly mock the args.func call
-    pytest.skip("Skipping test_main_help due to mocking issues")
