--- caylent-devcontainer-cli/tests/unit/test_code_more_coverage.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python3
-import os
-import sys
-
-# Add the parent directory to the path so we can import the CLI module
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
