--- caylent-devcontainer-cli/tests/unit/test_setup_interactive_version.py
@@ -1,166 +1,0 @@
-#!/usr/bin/env python3
-import json
-import os
-import sys
-from unittest.mock import mock_open, patch
-
-# Add the parent directory to the path so we can import the CLI module
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
-
-from caylent_devcontainer_cli.commands.setup_interactive import load_template_from_file
-
-
-def test_upgrade_template_with_container_env():
-    """Test upgrading a template that already has containerEnv."""
-    mock_template_data = {
-        "containerEnv": {"AWS_CONFIG_ENABLED": "true", "DEFAULT_GIT_BRANCH": "main"},
-        "aws_profile_map": {"default": {"region": "us-west-2"}},
-        "cli_version": "1.0.0",
-    }
-
-    expected_version = "2.0.0"
-
-    # We need to patch the upgrade_template function directly to use our version
-    with patch("caylent_devcontainer_cli.commands.setup_interactive.upgrade_template") as mock_upgrade:
-        # Set up the mock to call the real function but with our patched version
-        def side_effect(template_data):
-            with patch("caylent_devcontainer_cli.__version__", expected_version):
-                return {
-                    "containerEnv": template_data["containerEnv"],
-                    "aws_profile_map": template_data["aws_profile_map"],
-                    "cli_version": expected_version,
-                }
-
-        mock_upgrade.side_effect = side_effect
-
-        # Call the function through our mock
-        result = mock_upgrade(mock_template_data)
-
-    assert result["cli_version"] == expected_version
-    assert result["containerEnv"] == mock_template_data["containerEnv"]
-    assert result["aws_profile_map"] == mock_template_data["aws_profile_map"]
-
-
-def test_upgrade_template_with_env_values():
-    """Test upgrading a template that has env_values (old format)."""
-    mock_template_data = {
-        "env_values": {"AWS_CONFIG_ENABLED": "true", "DEFAULT_GIT_BRANCH": "main"},
-        "aws_profile_map": {"default": {"region": "us-west-2"}},
-        "cli_version": "1.0.0",
-    }
-
-    expected_version = "2.0.0"
-
-    # We need to patch the upgrade_template function directly to use our version
-    with patch("caylent_devcontainer_cli.commands.setup_interactive.upgrade_template") as mock_upgrade:
-        # Set up the mock to call the real function but with our patched version
-        def side_effect(template_data):
-            with patch("caylent_devcontainer_cli.__version__", expected_version):
-                return {
-                    "containerEnv": template_data["env_values"],
-                    "aws_profile_map": template_data["aws_profile_map"],
-                    "cli_version": expected_version,
-                }
-
-        mock_upgrade.side_effect = side_effect
-
-        # Call the function through our mock
-        result = mock_upgrade(mock_template_data)
-
-    assert result["cli_version"] == expected_version
-    assert result["containerEnv"] == mock_template_data["env_values"]
-    assert result["aws_profile_map"] == mock_template_data["aws_profile_map"]
-
-
-def test_upgrade_template_without_env_values():
-    """Test upgrading a template that has no env values."""
-    mock_template_data = {
-        "cli_version": "1.0.0",
-    }
-
-    expected_version = "2.0.0"
-    expected_env = {"AWS_CONFIG_ENABLED": "false"}
-
-    # We need to patch the upgrade_template function directly to use our version
-    with patch("caylent_devcontainer_cli.commands.setup_interactive.upgrade_template") as mock_upgrade:
-        # Set up the mock to call the real function but with our patched version
-        def side_effect(template_data):
-            return {
-                "containerEnv": expected_env,
-                "aws_profile_map": {},
-                "cli_version": expected_version,
-            }
-
-        mock_upgrade.side_effect = side_effect
-
-        # Call the function through our mock
-        result = mock_upgrade(mock_template_data)
-
-    assert result["cli_version"] == expected_version
-    assert result["containerEnv"] == expected_env
-    assert result["aws_profile_map"] == {}
-
-
-def test_upgrade_template_with_aws_enabled_no_profile():
-    """Test upgrading a template with AWS enabled but no profile map."""
-    mock_template_data = {
-        "containerEnv": {"AWS_CONFIG_ENABLED": "true", "DEFAULT_GIT_BRANCH": "main"},
-        "cli_version": "1.0.0",
-    }
-
-    expected_version = "2.0.0"
-    expected_aws_profile = {"default": {"region": "us-west-2"}}
-
-    # We need to patch the upgrade_template function directly to use our version
-    with patch("caylent_devcontainer_cli.commands.setup_interactive.upgrade_template") as mock_upgrade:
-        # Set up the mock to call the real function but with our patched version
-        def side_effect(template_data):
-            return {
-                "containerEnv": template_data["containerEnv"],
-                "aws_profile_map": expected_aws_profile,
-                "cli_version": expected_version,
-            }
-
-        mock_upgrade.side_effect = side_effect
-
-        # Call the function through our mock
-        result = mock_upgrade(mock_template_data)
-
-    assert result["cli_version"] == expected_version
-    assert result["containerEnv"] == mock_template_data["containerEnv"]
-    assert result["aws_profile_map"] == expected_aws_profile
-
-
-def test_load_template_from_file_with_version_parsing_error():
-    """Test loading a template with an invalid version string."""
-    mock_template_data = {
-        "containerEnv": {"AWS_CONFIG_ENABLED": "true"},
-        "cli_version": "invalid-version",
-    }
-
-    with patch("os.path.exists", return_value=True), patch(
-        "builtins.open", mock_open(read_data=json.dumps(mock_template_data))
-    ), patch("json.load", return_value=mock_template_data):
-        result = load_template_from_file("test-template")
-
-    assert result == mock_template_data
-
-
-def test_load_template_from_file_without_version():
-    """Test loading a template without version information."""
-    mock_template_data = {
-        "containerEnv": {"AWS_CONFIG_ENABLED": "true"},
-    }
-    expected_version = "2.0.0"
-
-    # Create a mock function for load_template_from_file that returns our expected result
-    with patch("caylent_devcontainer_cli.commands.setup_interactive.load_template_from_file") as mock_load:
-        # Set up the mock to return a result with our expected version
-        mock_result = mock_template_data.copy()
-        mock_result["cli_version"] = expected_version
-        mock_load.return_value = mock_result
-
-        # Call the function through our mock
-        result = mock_load("test-template")
-
-    assert result["cli_version"] == expected_version
