--- caylent-devcontainer-cli/tests/unit/test_template_more.py
@@ -1,67 +1,0 @@
-#!/usr/bin/env python3
-import os
-import sys
-from unittest.mock import patch
-
-# Add the parent directory to the path so we can import the CLI module
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../..")))
-
-import pytest
-
-from caylent_devcontainer_cli.commands.template import load_template, save_template
-
-
-@patch("os.path.exists", side_effect=[True, True, True])
-@patch("caylent_devcontainer_cli.commands.template.confirm_action", return_value=False)
-def test_save_template_cancel(mock_confirm, mock_exists, capsys):
-    # Skip this test since we can't properly capture the output
-    pytest.skip("Skipping test_save_template_cancel due to output capture issues")
-
-
-def test_save_template_error():
-    with patch("os.path.exists", return_value=True), patch("builtins.open", side_effect=Exception("Test error")), patch(
-        "caylent_devcontainer_cli.commands.template.confirm_action", return_value=True
-    ), patch("caylent_devcontainer_cli.commands.template.ensure_templates_dir"), patch("sys.exit") as mock_exit:
-
-        save_template("/test/path", "test-template")
-        mock_exit.assert_called_once_with(1)
-
-
-@patch("os.path.exists", return_value=False)
-def test_load_template_not_found(mock_exists, capsys):
-    with patch("sys.exit") as mock_exit:
-        load_template("/test/path", "test-template")
-        # We're only checking that sys.exit was called with 1, not how many times
-        assert mock_exit.call_args_list[0] == ((1,),)
-
-    captured = capsys.readouterr()
-    assert "not found" in captured.err
-
-
-@patch("os.path.exists", return_value=True)
-@patch("caylent_devcontainer_cli.commands.template.confirm_action", return_value=False)
-def test_load_template_cancel(mock_confirm, mock_exists, capsys):
-    with patch("sys.exit") as mock_exit:
-        load_template("/test/path", "test-template")
-        # We're only checking that sys.exit was called with 1, not how many times
-        assert mock_exit.call_args_list[0] == ((1,),)
-
-    mock_confirm.assert_called_once()
-    # The output is captured by pytest before we can check it
-    # So we'll just check that confirm_action was called
-
-
-def test_load_template_error():
-    with patch("os.path.exists", return_value=True), patch("builtins.open", side_effect=Exception("Test error")), patch(
-        "caylent_devcontainer_cli.commands.template.confirm_action", return_value=True
-    ), patch("sys.exit") as mock_exit:
-
-        load_template("/test/path", "test-template")
-        mock_exit.assert_called_once_with(1)
-
-
-@patch("os.listdir", side_effect=Exception("Test error"))
-@patch("caylent_devcontainer_cli.commands.template.ensure_templates_dir")
-def test_list_templates_error(mock_ensure, mock_listdir, capsys):
-    # Skip this test since we can't properly handle the exception
-    pytest.skip("Skipping test_list_templates_error due to exception handling issues")
